# src/generated/swagger.json
{
  "openapi": "3.0.0",
  "components": {
...
    "schemas": {
      "ProblemDetails": {
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "type": "object",
        "additionalProperties": false
      },
      "UUID": {
        "type": "string",
        "example": "7312cc99-f99f-445e-a939-eb66c0c6724c",
        "format": "uuid",
        "description": "Stringified UUIDv4.\nSee [RFC 4112](https://tools.ietf.org/html/rfc4122).",
        "pattern": "[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-4[0-9A-Fa-f]{3}-[89ABab][0-9A-Fa-f]{3}-[0-9A-Fa-f]{12}"
      },
...
      "Inventory": {
        "description": "Inventory of adoption status to quantities.",
        "properties": {
          "Adopted": {
            "type": "integer",
            "format": "int32",
            "description": "The number of pets that were adopted.",
            "example": 3
          },
          "Available": {
            "type": "integer",
            "format": "int32",
            "description": "The number of pets that are available for adoption.",
            "example": 1
          },
          "Pending": {
            "type": "integer",
            "format": "int32",
            "description": "The number of pets that have a pending adoption status.",
            "example": 2
          }
        },
        "required": ["Adopted", "Available", "Pending"],
        "type": "object",
        "additionalProperties": false
      },
      "AdoptionRequestStatus": {
        "description": "The adoption request's status.",
        "enum": ["Approved", "Pending", "Denied"],
        "type": "string"
      },
      "AdoptionRequest": {
        "description": "Adoption request information.",
        "properties": {
          "id": {
            "description": "The adoption's ID.",
            "example": "67120cf3-1434-44be-b660-b02df64db677",
            "allOf": [
              {
                "$ref": "#/components/schemas/UUID"
              }
            ]
          },
          "petId": {
            "description": "The adoptee's ID.",
            "example": "90dbbed9-bd3d-40ae-ad1c-86602844d4c1",
            "allOf": [
              {
                "$ref": "#/components/schemas/UUID"
              }
            ]
          },
          "dateOfSubmission": {
            "type": "string",
            "format": "date-time",
            "description": "The date of submission of the adoption request.",
            "example": "2024-08-25T00:00:00.000Z"
          },
          "status": {
            "description": "The adoption request status.",
            "example": "Pending",
            "allOf": [
              {
                "$ref": "#/components/schemas/AdoptionRequestStatus"
              }
            ]
          }
        },
        "required": ["id", "petId", "dateOfSubmission", "status"],
        "type": "object",
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "api_key": {
        "description": "The TSOA Demo Pet Store API Key is:\n\n`simple-pet-token`",
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "info": {
    "title": "TSOA Demo Pet Store - OpenAPI 3.0",
    "version": "1.0.0",
    "description": "Pet Store Server OpenAPI 3.0 specification.\n\nSome useful links:\n- [The TSOA Demo Pet Store repository](https://github.com/CloudNStoyan/tsoa-project-demo)\n- [The source API definition for the Pet Store](https://github.com/CloudNStoyan/tsoa-project-demo/blob/main/src/generated/swagger.yaml)",
    "contact": {}
  },
  "paths": {
...
    "/store/inventory": {
      "get": {
        "operationId": "GetInventory",
        "responses": {
          "200": {
            "description": "Successful retrieval of inventory.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Inventory"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "status": 401,
                      "title": "Unauthorized",
                      "detail": "Access denied!"
                    }
                  }
                }
              }
            }
          }
        },
        "description": "Returns a map of adoption status to quantities.",
        "summary": "Returns pet inventories by adoption status.",
        "tags": ["Store"],
        "security": [
          {
            "api_key": []
          }
        ],
        "parameters": []
      }
    },
    "/store/adopt": {
      "post": {
        "operationId": "AdoptPet",
        "responses": {
          "200": {
            "description": "Successful creation of adoption request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdoptionRequest"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "status": 401,
                      "title": "Unauthorized",
                      "detail": "Access denied!"
                    }
                  }
                }
              }
            }
          }
        },
        "description": "Place an adoption request for a pet.",
        "summary": "Request an adoption of a pet.",
        "tags": ["Store"],
        "security": [
          {
            "api_key": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "description": "The adoption request.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdoptionRequest",
                "description": "The adoption request."
              }
            }
          }
        }
      }
    },
    "/store/adopt/{requestId}": {
      "get": {
        "operationId": "GetAdoptRequestById",
        "responses": {
          "200": {
            "description": "Successful retrieval of adoption request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdoptionRequest"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "status": 401,
                      "title": "Unauthorized",
                      "detail": "Access denied!"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "status": 404,
                      "title": "Not Found",
                      "detail": "Adoption request not found!"
                    }
                  }
                }
              }
            }
          }
        },
        "description": "Find adoption request by ID.",
        "summary": "Find adoption request by ID.",
        "tags": ["Store"],
        "security": [
          {
            "api_key": []
          }
        ],
        "parameters": [
          {
            "description": "The adoption request's ID.",
            "in": "path",
            "name": "requestId",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UUID"
            }
          }
        ]
      },
      "delete": {
        "operationId": "DeleteAdoptRequestById",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "status": 401,
                      "title": "Unauthorized",
                      "detail": "Access denied!"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "status": 404,
                      "title": "Not Found",
                      "detail": "Adoption request not found!"
                    }
                  }
                }
              }
            }
          }
        },
        "description": "Delete adoption request by ID.",
        "summary": "Delete adoption request by ID.",
        "tags": ["Store"],
        "security": [
          {
            "api_key": []
          }
        ],
        "parameters": [
          {
            "description": "The adoption request's ID.",
            "in": "path",
            "name": "requestId",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UUID"
            }
          }
        ]
      }
    }
  },
  "servers": [],
...
}